% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExpectedMatrix.R
\name{calcExpectedMatrix}
\alias{calcExpectedMatrix}
\title{Calculate expected matrix}
\usage{
calcExpectedMatrix(hypot, ztrans_coef)
}
\arguments{
\item{hypot}{a single modularity hypothesis}

\item{ztrans_coef}{a vector of z-transformed correlations for each module, 
starting with a background correlation}
}
\description{
Calculates the expected correlation matrix for a given modularity hypothesis 
using background and within-modules z-transformed correlations
}
