% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExpectedMatrixFactors.R
\name{calcExpectedMatrixFactors}
\alias{calcExpectedMatrixFactors}
\title{Calculate expected matrix using modularity Factors}
\usage{
calcExpectedMatrixFactors(hypot, coef, c2c = TRUE)
}
\arguments{
\item{hypot}{a single modularity hypothesis}

\item{coef}{a vector of coefficients for each module, 
starting with a background correlation}

\item{c2c}{Logical. If TRUE, transform resulting matrix to correlation matrix.}
}
\description{
Calculates the expected correlation matrix for a given modularity hypothesis 
using background and within-modules coefficients
}
