% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcZTransCoef.R
\name{calcZTransCoef}
\alias{calcZTransCoef}
\title{Calculate z-tranform of module correlations using a correlation matrix}
\usage{
calcZTransCoef(hypot, x, nneg)
}
\arguments{
\item{hypot}{a single modularity hypothesis}

\item{x}{a correlation matrix}

\item{nneg}{logical. If TRUE, module coeficients can only be positive}
}
\description{
calculates the z-tranform for a given correlation matrix and modularity hypothesis
}
