% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFactorsML.R
\name{fitFactorsML}
\alias{fitFactorsML}
\title{Calculate the z-tranform of module correlations using individual values}
\usage{
fitFactorsML(data, hypot, nneg)
}
\arguments{
\item{data}{individual measurements or residuals}

\item{hypot}{a single modularity hypothesis}
}
\description{
Calculates the z-tranform for a given population and modularity hypothesis
}
