% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModuleCoef.R
\name{fitModuleCoef}
\alias{fitModuleCoef}
\title{Calculate the coefficients of module correlations using individual values}
\usage{
fitModuleCoef(data, hypot, nneg, factors)
}
\arguments{
\item{data}{individual measurements or residuals}

\item{hypot}{a single modularity hypothesis}

\item{nneg}{Logical. If TRUE, coeficients are non-negative?}

\item{factors}{Use factors or z-transform additive correlations}
}
\description{
Calculates the z-transform for a given population and modularity hypothesis
}
