% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_hypot.R
\name{hclustHypot}
\alias{hclustHypot}
\title{Hierarchical clustering with hclust}
\usage{
hclustHypot(x, ...)
}
\arguments{
\item{x}{correlation matrix}

\item{...}{aditional arguments to hclust}
}
\description{
Generate modularity hypothesis using hierarchical clustering via hclust
}
