% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamdar.R
\docType{data}
\name{toadCor}
\alias{toadCor}
\title{Example correlation matrix for a toad skull}
\format{
Matrix
}
\usage{
data(toadCor)
}
\description{
Correlation matrix from "High evolutionary constraints limited adaptive responses to past climate changes in toad skulls" Monique Nouailhetas Simon, Fabio Andrade Machado, Gabriel Marroig Proc. R. Soc. B 2016 283 20161783; DOI: 10.1098/rspb.2016.1783. Published 26 October 2016
}
\keyword{datasets}
