% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamdar.R
\docType{data}
\name{toadHypo}
\alias{toadHypo}
\title{Example modularity hypothesis for the toad skull}
\format{
Matrix
}
\usage{
data(toadHypo)
}
\description{
List of modularity hypothesis matrices for the toad skull
}
\keyword{datasets}
