% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toHypotMatrix.R
\name{toHypotMatrix}
\alias{toHypotMatrix}
\title{to Hypot matrix}
\usage{
toHypotMatrix(membership)
}
\arguments{
\item{membership}{vector of module membership.}
}
\description{
Transform membership vector to a matrix of modularity hypothesis
}
