% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toMembership.R
\name{toMembership}
\alias{toMembership}
\title{to Membership vector}
\usage{
toMembership(hypot, membershipAsFactor = FALSE, noModule = NULL)
}
\arguments{
\item{hypot}{matrix of modularity hypothesis. If position [i,j] is one, the trait i belong to module j, and if it is zero, trait i does not belong to module j. Modules can be overlapping.}

\item{membershipAsFactor}{Logical. Transform membership vector to factor?}

\item{noModule}{String. Module to assing traits that don't belong to any modules. Leave empty to use NA}
}
\description{
Transform a matrix of modularity hypothesis into a mebership vector
}
