% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamdaMLE.R
\name{YamdaMLE}
\alias{YamdaMLE}
\title{Yet another modularity detection algorith in R}
\usage{
YamdaMLE(data, hypot_list, nneg = TRUE)
}
\arguments{
\item{data}{individual measurements or residuals}

\item{hypot_list}{list of matrices describing modularity hypothesis. Each element in the list should be one matrix, and each column in these matrices represents a module, and each row a trait. If position [i,j] is one, the trait i belong to module j, and if it is zero, trait i does not belong to module j. Modules can be overlapping.}

\item{nneg}{If true, belonging to the same module can only increase correlations, not decrease them.}
}
\description{
Uses hierarchical partitions and MLE to rank modularity hypothesis using individual measurements.
}
\examples{
library(mvtnorm)
modules = matrix(c(rep(c(1, 0, 0), each = 5),
                   rep(c(0, 1, 0), each = 5),
                   rep(c(0, 0, 1), each = 5)), 15)
modules_ztrans = c(0.3, 0.5, .3, 1.7)
mod.cor = calcExpectedMatrix(modules, modules_ztrans)
sds = runif(15, 1, 10)
mod_cov = outer(sds, sds) * mod.cor
pop = rmvnorm(50, sigma = mod_cov)
YamdaMLE(pop, list(modules, modules[,-1]), TRUE)[[2]]
}
